move more stuff out of __init__
