+++ joplin/base/models/__init__.py
@@ -22,99 +22,21 @@ from base import blocks as custom_blocks
 from base import forms as custom_forms
 
 from .translated_image import TranslatedImage
+from .contact import Contact
+from .location import Location
 
 from .janis_page import JanisPage, JanisBasePage
 from .home_page import HomePage
+from .theme import Theme
 from .topic_collection_page import TopicCollectionPage
 from .topic_page import TopicPage
-from .service_page import ServicePage
+from .service_page import ServicePage, ServicePageTopic, ServicePageContact, ServicePageRelatedDepartments
 from .information_page import InformationPage
-from .department_page import DepartmentPage
+from .department_page import DepartmentPage, DepartmentPageDirector
 
 WYSIWYG_GENERAL = ['h1', 'h2', 'h3', 'h4', 'bold', 'link', 'ul', 'ol', 'code']
 DEFAULT_MAX_LENGTH = 255
 SHORT_DESCRIPTION_LENGTH = 300
-
-@register_snippet
-class ThreeOneOne(ClusterableModel):
-    title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    url = models.URLField()
-
-    def __str__(self):
-        return self.title
-
-
-class ProcessPage(JanisPage):
-    janis_url_page_type = "processes"
-
-    department = models.ForeignKey(
-        'base.DepartmentPage',
-        on_delete=models.PROTECT,
-        verbose_name='Select a Department',
-        blank=True,
-        null=True,
-    )
-
-    description = models.TextField(blank=True)
-    image = models.ForeignKey(TranslatedImage, null=True, blank=True, on_delete=models.SET_NULL, related_name='+')
-    # TODO: Add images array field
-
-    base_form_class = custom_forms.ProcessPageForm
-
-    content_panels = [
-        InlinePanel('topics', label='Topics'),
-        FieldPanel('department'),
-        FieldPanel('description'),
-        ImageChooserPanel('image'),
-        InlinePanel('contacts', label='Contacts'),
-        InlinePanel('process_steps', label="Process steps"),
-    ]
-
-class DepartmentPageDirector(Orderable):
-    page = ParentalKey(DepartmentPage, related_name='department_directors')
-    name = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    title = models.CharField(max_length=DEFAULT_MAX_LENGTH, default='Director')
-    photo = models.ForeignKey(TranslatedImage, null=True, blank=True, on_delete=models.SET_NULL, related_name='+')
-    about = models.TextField(blank=True)
-
-    panels = [
-        FieldPanel('name'),
-        FieldPanel('title'),
-        ImageChooserPanel('photo'),
-        FieldPanel('about'),
-    ]
-
-class ProcessPageStep(Orderable):
-    page = ParentalKey(ProcessPage, related_name='process_steps')
-    title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    short_title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    link_title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    description = models.TextField(blank=True)
-    image = models.ForeignKey(TranslatedImage, null=True, blank=True, on_delete=models.SET_NULL, related_name='+')
-    overview_steps = RichTextField(features=WYSIWYG_GENERAL, verbose_name='Write out the steps a resident needs to take to use the service', blank=True)
-    detailed_content = RichTextField(features=WYSIWYG_GENERAL, verbose_name='Write any detailed content describing the process', blank=True)
-    quote = models.TextField(blank=True)
-
-    panels = [
-        FieldPanel('title'),
-        FieldPanel('short_title'),
-        FieldPanel('link_title'),
-        FieldPanel('description'),
-        ImageChooserPanel('image'),
-        FieldPanel('overview_steps'),
-        FieldPanel('detailed_content'),
-        FieldPanel('quote'),
-    ]
-
-
-@register_snippet
-class Theme(ClusterableModel):
-    slug = models.SlugField()
-    text = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    description = models.TextField()
-
-    def __str__(self):
-        return self.text
 
 
 @register_snippet
@@ -134,41 +56,6 @@ class Map(ClusterableModel):
         data['location'].pop('pk')
 
         return data
-
-
-@register_snippet
-class Location(ClusterableModel):
-    TX = 'TX'
-    STATE_CHOICES = (
-        (TX, 'Texas'),
-    )
-
-    USA = 'United States'
-    COUNTRY_CHOICES = (
-        (USA, 'United States'),
-    )
-
-    name = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    street = models.TextField()
-    city = models.CharField(max_length=DEFAULT_MAX_LENGTH, default='Austin')
-    state = models.CharField(max_length=2, choices=STATE_CHOICES, default=TX)
-    country = models.CharField(max_length=100, choices=COUNTRY_CHOICES, default=USA)
-    zip = models.CharField(max_length=50)
-
-    panels = [
-        FieldPanel('name'),
-        MultiFieldPanel(children=[
-            FieldPanel('street'),
-            FieldPanel('city', classname='col5'),
-            FieldPanel('state', classname='col4'),
-            FieldPanel('zip', classname='col2'),
-            FieldPanel('country', classname='col5'),
-        ], heading='Location'),
-    ]
-
-    def __str__(self):
-        return self.name
-
 
 class DayAndDuration(ClusterableModel):
     MONDAY = 'Monday'
@@ -201,89 +88,11 @@ class DayAndDuration(ClusterableModel):
     def __str__(self):
         return f'{self.day_of_week} {self.start_time} - {self.end_time}'
 
-
-@register_snippet
-class Contact(ClusterableModel):
-    name = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    email = models.EmailField()
-    phone = models.CharField(max_length=DEFAULT_MAX_LENGTH)
-    location = models.ForeignKey(Location, null=True, blank=True, related_name='+', on_delete=models.SET_NULL)
-
-    social_media = StreamField(
-        [
-            ('url', URLBlock(
-                label='Social media url'
-            ))
-        ],
-        verbose_name='Links to any social media pages',
-        help_text='For example: https://www.facebook.com/atxpoliceoversight/',
-        blank=True
-    )
-
-    panels = [
-        FieldPanel('name'),
-        FieldPanel('email'),
-        FieldPanel('phone'),
-        SnippetChooserPanel('location'),
-        InlinePanel('hours', label='Hours'),
-        StreamFieldPanel('social_media'),
-    ]
-
-    def __str__(self):
-        return self.name
-
-
 class ContactDayAndDuration(Orderable, DayAndDuration):
     contact = ParentalKey(Contact, related_name='hours')
 
     content_panels = [
         SnippetChooserPanel('day_and_duration'),
-    ]
-
-
-class ProcessPageContact(ClusterableModel):
-    process = ParentalKey(ProcessPage, related_name='contacts')
-    contact = models.ForeignKey(Contact, related_name='+', on_delete=models.CASCADE)
-
-    panels = [
-        SnippetChooserPanel('contact'),
-    ]
-
-    def __str__(self):
-        return self.contact.name
-
-class ProcessPageTopic(ClusterableModel):
-    page = ParentalKey(ProcessPage, related_name='topics')
-    topic = models.ForeignKey('base.TopicPage',  verbose_name='Select a Topic', related_name='+', on_delete=models.CASCADE)
-
-    panels = [
-        PageChooserPanel('topic'),
-    ]
-
-    def __str__(self):
-        return self.topic.text
-
-class ServicePageContact(ClusterableModel):
-    page = ParentalKey(ServicePage, related_name='contacts')
-    contact = models.ForeignKey(Contact, related_name='+', on_delete=models.CASCADE)
-
-    panels = [
-        SnippetChooserPanel('contact'),
-    ]
-
-    def __str__(self):
-        return self.contact.name
-
-class ServicePageRelatedDepartments(ClusterableModel):
-    page = ParentalKey(ServicePage, related_name='related_departments', default=None)
-    related_department = models.ForeignKey(
-        "base.departmentPage",
-        on_delete=models.PROTECT,
-    )
-
-    panels = [
-        # Use a SnippetChooserPanel because blog.BlogAuthor is registered as a snippet
-        PageChooserPanel("related_department"),
     ]
 
 class InformationPageRelatedDepartments(ClusterableModel):
@@ -320,22 +129,6 @@ class TopicPageTopicCollection(ClusterableModel):
     def __str__(self):
         return self.topiccollection.text
 
-class ServicePageTopic(ClusterableModel):
-    page = ParentalKey(ServicePage, related_name='topics')
-    topic = models.ForeignKey('base.TopicPage',  verbose_name='Select a Topic', related_name='+', on_delete=models.CASCADE)
-    toplink = models.BooleanField(default=False, verbose_name='Make this service a top link for this topic')
-
-    panels = [
-        MultiFieldPanel(
-            [
-                PageChooserPanel('topic'),
-                FieldPanel('toplink'),
-            ]
-        ),
-    ]
-
-    def __str__(self):
-        return self.topic.title
 
 class InformationPageContact(ClusterableModel):
     page = ParentalKey(InformationPage, related_name='contacts')
@@ -375,6 +168,89 @@ class DepartmentPageContact(ClusterableModel):
 
     def __str__(self):
         return self.contact.name
+
+
+
+# TODO: Remove everything below this comment
+
+@register_snippet
+class ThreeOneOne(ClusterableModel):
+    title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
+    url = models.URLField()
+
+    def __str__(self):
+        return self.title
+
+
+class ProcessPage(JanisPage):
+    janis_url_page_type = "processes"
+
+    department = models.ForeignKey(
+        'base.DepartmentPage',
+        on_delete=models.PROTECT,
+        verbose_name='Select a Department',
+        blank=True,
+        null=True,
+    )
+
+    description = models.TextField(blank=True)
+    image = models.ForeignKey(TranslatedImage, null=True, blank=True, on_delete=models.SET_NULL, related_name='+')
+    # TODO: Add images array field
+
+    base_form_class = custom_forms.ProcessPageForm
+
+    content_panels = [
+        InlinePanel('topics', label='Topics'),
+        FieldPanel('department'),
+        FieldPanel('description'),
+        ImageChooserPanel('image'),
+        InlinePanel('contacts', label='Contacts'),
+        InlinePanel('process_steps', label="Process steps"),
+    ]
+
+class ProcessPageStep(Orderable):
+    page = ParentalKey(ProcessPage, related_name='process_steps')
+    title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
+    short_title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
+    link_title = models.CharField(max_length=DEFAULT_MAX_LENGTH)
+    description = models.TextField(blank=True)
+    image = models.ForeignKey(TranslatedImage, null=True, blank=True, on_delete=models.SET_NULL, related_name='+')
+    overview_steps = RichTextField(features=WYSIWYG_GENERAL, verbose_name='Write out the steps a resident needs to take to use the service', blank=True)
+    detailed_content = RichTextField(features=WYSIWYG_GENERAL, verbose_name='Write any detailed content describing the process', blank=True)
+    quote = models.TextField(blank=True)
+
+    panels = [
+        FieldPanel('title'),
+        FieldPanel('short_title'),
+        FieldPanel('link_title'),
+        FieldPanel('description'),
+        ImageChooserPanel('image'),
+        FieldPanel('overview_steps'),
+        FieldPanel('detailed_content'),
+        FieldPanel('quote'),
+    ]
+
+class ProcessPageContact(ClusterableModel):
+    process = ParentalKey(ProcessPage, related_name='contacts')
+    contact = models.ForeignKey(Contact, related_name='+', on_delete=models.CASCADE)
+
+    panels = [
+        SnippetChooserPanel('contact'),
+    ]
+
+    def __str__(self):
+        return self.contact.name
+
+class ProcessPageTopic(ClusterableModel):
+    page = ParentalKey(ProcessPage, related_name='topics')
+    topic = models.ForeignKey('base.TopicPage',  verbose_name='Select a Topic', related_name='+', on_delete=models.CASCADE)
+
+    panels = [
+        PageChooserPanel('topic'),
+    ]
+
+    def __str__(self):
+        return self.topic.text
 
 @register_snippet
 class Department(ClusterableModel):
