+++ joplin/base/models/constants.py
@@ -1,2 +1,3 @@
 WYSIWYG_GENERAL = ['h1', 'h2', 'h3', 'h4', 'bold', 'link', 'ul', 'ol', 'code']
 SHORT_DESCRIPTION_LENGTH = 300
+DEFAULT_MAX_LENGTH = 255
